+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -117,6 +117,11 @@ Improvements to clang-tidy
 New checks
 ^^^^^^^^^^
 
+- New :doc:`BSCompatibility-non-void-function-return-void
+  <clang-tidy/checks/BSCompatibility/non-void-function-return-void>` check.
+
+  Check if a non-void function hasn't return statement.
+
 - New :doc:`bugprone-empty-catch
   <clang-tidy/checks/bugprone/empty-catch>` check.